Moving testing util to fixture.rs
