+++ contracts/dca/src/handlers/cancel_dca_order.rs
@@ -96,7 +96,7 @@ mod tests {
         Empty, Response,
     };
 
-    use super::super::deposit::test_util::mock_storage_valid_data;
+    use crate::fixture::fixture::mock_storage_valid_data;
     use crate::contract::execute;
 
     #[test]
