+++ contracts/dca/src/utils.rs
@@ -1,16 +1,38 @@
-use crate::state::{Config, CONFIG};
-use crate::{
-    error::ContractError,
-    get_token_allowance::get_token_allowance,
-    state::{DCA_ORDERS, USER_DCA_ORDERS},
-};
-
+use crate::error::ContractError;
 use astroport::asset::{Asset, AssetInfo};
-use astroport_dca::dca::{Balance, DcaInfo, WhitelistedTokens};
-use cosmwasm_std::{
-    attr, to_binary, BankMsg, Coin, CosmosMsg, DepsMut, MessageInfo, Response, WasmMsg,
-};
+use cosmwasm_std::{to_binary, BankMsg, Coin, CosmosMsg, MessageInfo, WasmMsg};
 use cw20::Cw20ExecuteMsg;
+
+use cosmwasm_std::{Addr, Deps, Env, StdResult, Uint128};
+use cw20::{AllowanceResponse, Cw20QueryMsg};
+
+/// ## Description
+/// Retrieves the allowed token allowance for the contract for a Cw20 token as a [`Uint128`].
+/// ## Arguments
+/// * `deps` - A [`Deps`] that contains the dependencies.
+///
+/// * `env` - The [`Env`] of the blockchain.
+///
+/// * `owner` - The user that holds the Cw20 token.
+///
+/// * `contract_address` - The address of the Cw20 token.
+pub fn get_token_allowance(
+    deps: &Deps,
+    env: &Env,
+    owner: &Addr,
+    contract_address: &Addr,
+) -> StdResult<Uint128> {
+    let allowance_response: AllowanceResponse = deps.querier.query_wasm_smart(
+        contract_address,
+        &Cw20QueryMsg::Allowance {
+            owner: owner.to_string(),
+            spender: env.contract.address.to_string(),
+        },
+    )?;
+
+    Ok(allowance_response.allowance)
+}
+
 /*
 
 // check deposit asset is in the Whitelist
@@ -52,4 +74,3 @@ pub fn build_send_message(info: MessageInfo, asset: Asset) -> Result<CosmosMsg, 
 
     Ok(message)
 }
-
